# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TextKit API",
    "version": "1.0.0",
    "description": "Pay-per-use text/content utility API for RapidAPI marketplace. Transform, analyze, and enhance text content with 9 powerful AI-powered endpoints.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "RapidAPIProxySecret": []
    },
    {
      "ApiKey": []
    },
    {
      "JwtCookie": []
    }
  ],
  "paths": {
...
    "/dashboard/studio/usage": {
      "get": {
        "summary": "Get Studio usage stats",
        "description": "Returns current user's tier, today's usage count, and rate limits for TextKit Studio interface",
        "tags": ["Dashboard"],
        "security": [
          {
            "JwtCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usage statistics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tier": {
                      "type": "string",
                      "enum": ["FREE", "BASIC", "PRO", "ULTRA", "MEGA"],
                      "description": "Current subscription tier"
                    },
                    "usedToday": {
                      "type": "integer",
                      "description": "Number of API calls made today"
                    },
                    "limitToday": {
                      "type": "integer",
                      "nullable": true,
                      "description": "Daily request limit (null for unlimited)"
                    },
                    "limitPerMinute": {
                      "type": "integer",
                      "description": "Per-minute request limit"
                    }
                  }
                },
                "example": {
                  "tier": "PRO",
                  "usedToday": 42,
                  "limitToday": 500,
                  "limitPerMinute": 30
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "RapidAPIProxySecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Proxy-Secret",
        "description": "RapidAPI proxy secret for production authentication"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Local API key for development authentication"
      },
      "JwtCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token cookie for web application authentication (used by TextKit Studio)"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "code": {
                  "type": "string",
                  "example": "AUTH_REQUIRED"
                },
                "details": {
                  "type": "string",
                  "example": "Provide x-rapidapi-proxy-secret or x-api-key header"
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}